Fixed category cards on homepage to redirect to correct pages, moved auth buttons (login/register) to the right side of navbar, edit button no longer shows up in component info for users that do not have the role ADMIN

+++ src/pages/ComponentTable.js
@@ -42,11 +42,14 @@ const ComponentList = ( {category} ) => {
     }
 
     const handleDelete = async (id) => {
-        try {
-            await axios.delete(`/components/${id}`);
-            fetchComponents();
-        } catch (error) {
-            console.error('Error trying to delete component: ', error)
+        if(window.confirm("Are you sure you want to delete this?")) {
+            try {
+                await axios.delete(`/components/${id}`);
+                fetchComponents();
+                alert('Deleted Successfuly!')
+            } catch (error) {
+                console.error('Error trying to delete component: ', error)
+            }
         }
     }
